requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pdfplumber==0.10.2
langchain==0.0.335
langchain-openai==0.0.2
langchain-community>=0.0.1
faiss-cpu==1.7.4
python-dotenv==1.0.0
websockets==12.0
markdown==3.5
openai>=1.6.1,<2.0.0
pydantic==2.4.2
numpy==1.26.1
tiktoken>=0.5.2,<0.6.0
